translator/requirements.txt
IndicTransTokenizer @ git+https://github.com/VarunGumma/IndicTransTokenizer
python-dotenv
torch
transformers